Warehouse bug fix and process validation, added actual request stock id for the intemized item

+++ src/main/java/com/boheco1/dev/integratedaccountingsystem/objects/StockDescription.java
@@ -3,13 +3,14 @@ package com.boheco1.dev.integratedaccountingsystem.objects;
 public class StockDescription {
     private String id;
     private String description;
-    private int quantity;
+    private double quantity;
 
-    public StockDescription(String id, String description, int quantity) {
+    public StockDescription(String id, String description, double quantity) {
         this.id = id;
         this.description = description;
         this.quantity = quantity;
     }
+
 
     public String getId() {
         return id;
@@ -27,11 +28,11 @@ public class StockDescription {
         this.description = description;
     }
 
-    public int getQuantity() {
+    public double getQuantity() {
         return quantity;
     }
 
-    public void setQuantity(int quantity) {
+    public void setQuantity(double quantity) {
         this.quantity = quantity;
     }
 }
